move detr collater
